Move form error normalizer to the Serializer component
