+++ src/Symfony/Component/Serializer/composer.json
@@ -29,6 +29,7 @@
         "symfony/dependency-injection": "^4.4|^5.0",
         "symfony/error-handler": "^4.4|^5.0",
         "symfony/filesystem": "^4.4|^5.0",
+        "symfony/form": "^4.4|^5.0",
         "symfony/http-foundation": "^4.4|^5.0",
         "symfony/http-kernel": "^4.4|^5.0",
         "symfony/mime": "^4.4|^5.0",
